+++ src/main/java/org/ilgcc/app/pdf/ApplicantEducationSchedulePreparer.java
@@ -10,6 +10,7 @@ import java.util.Optional;
 import org.ilgcc.app.utils.ActivitySchedules.HourlySchedule;
 import org.ilgcc.app.utils.ActivitySchedules.LocalTimeRange;
 import org.ilgcc.app.utils.DayOfWeekOption;
+import org.ilgcc.app.utils.SchedulePreparerUtility;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -18,7 +19,7 @@ public class ApplicantEducationSchedulePreparer implements SubmissionFieldPrepar
     public Map<String, SubmissionField> prepareSubmissionFields(Submission submission, PdfMap pdfMap) {
         var results = new HashMap<String, SubmissionField>();
 
-        Optional<HourlySchedule> educationSchedule = SchedulePreparer.getHourlySchedule(submission, "activitiesClass", "weeklySchedule[]");
+        Optional<HourlySchedule> educationSchedule = SchedulePreparerUtility.getHourlySchedule(submission, "activitiesClass", "weeklySchedule[]");
         if (educationSchedule.isEmpty()) {
             return results;
         }
@@ -27,7 +28,7 @@ public class ApplicantEducationSchedulePreparer implements SubmissionFieldPrepar
         for (var scheduleEntry : dailyScheduleMap.entrySet()) {
             DayOfWeekOption day = scheduleEntry.getKey();
             LocalTimeRange schedule = scheduleEntry.getValue();
-            results.putAll(SchedulePreparer.createSubmissionFieldsFromSchedule(schedule, day, "applicantEducationSchedule"));
+            results.putAll(SchedulePreparerUtility.createSubmissionFieldsFromSchedule(schedule, day, "applicantEducationSchedule"));
         }
         return results;
     }
